docs: remove out-of-date sample from README
See samples/ directory for maintained and tested samples
