+++ samples/snippets/requirements.txt
@@ -1,1 +1,1 @@
-google-cloud-bigquery-datatransfer==2.1.0
+google-cloud-bigquery-datatransfer==3.0.0
